Start splitting out menu
