+++ config/splat.pspeu.dra.yaml
@@ -53,6 +53,7 @@ segments:
       - [0xA710, c, dra_psp/A710] # O4
       - [0xBDE0, c, dra_psp/BDE0]
       - [0xC0B0, c, dra_psp/C0B0] # O4
+      - [0xE6A8, c, dra_psp/E6A8]
       - [0xEC60, c, dra_psp/EC60] # O4
       - [0xF4A0, c, dra_psp/F4A0]
       - [0xF4D0, c, dra_psp/F4D0]
