+++ src/dra_psp/C0B0.c
@@ -57,128 +57,3 @@ INCLUDE_ASM("dra_psp/psp/dra_psp/C0B0", func_800F298C);
 INCLUDE_ASM("dra_psp/psp/dra_psp/C0B0", func_psp_090EAF08);
 
 INCLUDE_ASM("dra_psp/psp/dra_psp/C0B0", func_psp_090EAFA8);
-
-bool CheckIfAllButtonsAreAssigned(void) {
-    s32 buf[8];
-    s32 i;
-    s32 bitMask_Assigned;
-    s32* miscPtr;
-
-    for (i = 1; i < 8; i++) {
-        buf[i] = 0;
-    }
-
-    for (i = 0; i < 7; i++) {
-        buf[g_Settings.buttonConfig[i] + 1] = 1;
-    }
-
-    // What is the purpose of this loop?
-    for (i = 0, miscPtr = &buf[1]; i < 7; miscPtr++, i++) {
-        if (*miscPtr == 0) {
-            break;
-        }
-    }
-
-    for (i = 0; i < 7; i++) {
-        g_Settings.buttonMask[i] = g_ButtonMask[g_Settings.buttonConfig[i]];
-    }
-
-    bitMask_Assigned = 0;
-    miscPtr = g_Settings.buttonConfig;
-    for (i = 1; i < 8; i++) {
-        bitMask_Assigned |= 1 << *miscPtr++;
-    }
-
-    return bitMask_Assigned == 0x7F;
-}
-
-INCLUDE_ASM("dra_psp/psp/dra_psp/C0B0", IsAlucart);
-
-extern s32 g_StatBuffTimers[16];
-
-void func_800F4994(void) {
-    s32* statsPtr = &g_Status.statsEquip[0];
-    s32 correctStonesEquipped;
-    s32 hourOfDay;
-    s32 i, j;
-    s32 statBonus;
-
-    for (i = 0; i < 4; i++, statsPtr++) {
-        *statsPtr = 0;
-    }
-
-    // Iterate through each Item Slot
-    for (i = 0; i < 5; i++) {
-        // Iterate through the 4 stats (STR, CON, INT, LCK)
-        for (j = 0; j < 4; j++) {
-            statBonus = g_AccessoryDefs[g_Status.equipment[HEAD_SLOT + i]]
-                            .statsBonus[j];
-            if (statBonus > 128) {
-                statBonus -= 256;
-            }
-            g_Status.statsEquip[j] += statBonus;
-        }
-    }
-    hourOfDay = g_Status.timerHours % 24;
-
-    // Hours of sunstone effectiveness
-    if (6 <= hourOfDay && hourOfDay < 18) {
-        // Sunstone check
-        correctStonesEquipped =
-            CheckEquipmentItemCount(ITEM_SUNSTONE, EQUIP_ACCESSORY);
-        statsPtr = &g_Status.statsEquip[0];
-        for (i = 0; i < 4; i++, statsPtr++) {
-            *statsPtr += correctStonesEquipped * 5;
-        }
-    } else {
-        // Moonstone check
-        correctStonesEquipped =
-            CheckEquipmentItemCount(ITEM_MOONSTONE, EQUIP_ACCESSORY);
-        statsPtr = &g_Status.statsEquip[0];
-        for (i = 0; i < 4; i++, statsPtr++) {
-            *statsPtr += correctStonesEquipped * 5;
-        }
-    }
-
-    if (g_StatBuffTimers[4]) {
-        g_Status.statsEquip[STAT_STR] += 20;
-    }
-    if (g_StatBuffTimers[3]) {
-        g_Status.statsEquip[STAT_INT] += 20;
-    }
-    if (g_StatBuffTimers[2]) {
-        g_Status.statsEquip[STAT_LCK] += 20;
-    }
-    if (g_Status.relics[RELIC_RIB_OF_VLAD] & 2) {
-        g_Status.statsEquip[STAT_CON] += 10;
-    }
-    if (g_Status.relics[RELIC_EYE_OF_VLAD] & 2) {
-        g_Status.statsEquip[STAT_LCK] += 10;
-    }
-    if (g_Status.relics[RELIC_TOOTH_OF_VLAD] & 2) {
-        g_Status.statsEquip[STAT_STR] += 10;
-    }
-    if (g_Status.relics[RELIC_RING_OF_VLAD] & 2) {
-        g_Status.statsEquip[STAT_INT] += 10;
-    }
-    if (IsAlucart() != false) {
-        g_Status.statsEquip[STAT_LCK] += 30;
-    }
-
-    for (i = 0; i < 4; i++) {
-        if (g_Status.statsEquip[i] > 99) {
-            g_Status.statsEquip[i] = 99;
-        }
-        g_Status.statsTotal[i] = g_Status.statsBase[i] + g_Status.statsEquip[i];
-    }
-
-    g_Status.statsTotal[1] =
-        (g_Status.statsEquip[1] * 8) + g_Status.statsBase[1];
-    g_Status.statsTotal[2] =
-        (g_Status.statsEquip[2] * 4) + g_Status.statsBase[2];
-    for (i = 0; i < 4; i++) {
-        if (g_Status.statsTotal[i] < 0) {
-            g_Status.statsTotal[i] = 0;
-        }
-    }
-}
